backend/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
duckduckgo-search>=5.0.0
google-adk[extensions]>=1.22.0
docker>=7.0.0
asyncpg>=0.29.0
tweepy>=4.16.0

